Edit CommentService

+++ src/main/java/br/com/projectblog/services/CommentService.java
@@ -6,16 +6,23 @@ import java.util.UUID;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
+import br.com.projectblog.domains.Role;
 import br.com.projectblog.dtos.CommentDTO;
 import br.com.projectblog.dtos.PostDTO;
+import br.com.projectblog.dtos.UserDTO;
 import br.com.projectblog.dtos.requests.CommentRequestDTO;
+import br.com.projectblog.exceptions.BusinessException;
 import br.com.projectblog.exceptions.ResourceNotFoundException;
 import br.com.projectblog.mappers.CommentMapper;
 import br.com.projectblog.mappers.PostMapper;
 import br.com.projectblog.models.Comment;
+import br.com.projectblog.models.Post;
+import br.com.projectblog.models.User;
 import br.com.projectblog.repositories.CommentRepository;
+import br.com.projectblog.utils.UserUtils;
 import br.com.projectblog.validations.groups.OnCreate;
 import br.com.projectblog.validations.groups.OnUpdate;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -38,19 +45,23 @@ public class CommentService {
 	
 	
 	@Validated({ OnCreate.class })
-	public CommentDTO insertCommentInPost(CommentRequestDTO dto) {
+	public CommentDTO insertCommentInPost(@Valid CommentRequestDTO dto) {
 		log.info("Requisição para criação do comment: \n {}", dto);
 		
-		PostDTO postDTO = this.postService.findById(UUID.fromString(dto.getPostId())).orElseThrow(() -> 
+		Post post = this.postService.findByIdPost(UUID.fromString(dto.getPostId())).orElseThrow(() -> 
 			new ResourceNotFoundException(String.format("Post pelo identificador %s não encontrado para ser atualizado.", dto.getPostId())
 		));
 		
-		this.userService.findById(UUID.fromString(dto.getUserId())).orElseThrow(() -> 
+		User user = this.userService.findByIdUser(UUID.fromString(dto.getUserId())).orElseThrow(() -> 
 			new ResourceNotFoundException(String.format("Usuário pelo identificador %s não encontrado.", dto.getUserId())
 		));
 		
+		if(!UserUtils.getUsernameLogado().equals(user.getUsername()))
+			throw new BusinessException("Não é possível inserir comentário no post utilizando outro usuário.");
+		
 		Comment entity = this.mapper.commentRequestDTOToComment(dto);
-		entity.setPost(this.postMapper.postDTOToPost(postDTO));
+		entity.setPost(post);
+		entity.setUser(user);
 		entity.setDateCreate(LocalDateTime.now());
 		entity.setDateUpdate(LocalDateTime.now());
 		
@@ -63,7 +74,7 @@ public class CommentService {
 	
 	
 	@Validated({ OnUpdate.class })
-	public CommentDTO updateCommentInPost(CommentRequestDTO dto) {
+	public CommentDTO updateCommentInPost(@Valid CommentRequestDTO dto) {
 		log.info("Requisição para atualização do comment: \n {}", dto);
 		
 		Comment actualEntity = this.repository.findById(UUID.fromString(dto.getCommentId())).orElseThrow(() -> 
@@ -74,10 +85,13 @@ public class CommentService {
 			new ResourceNotFoundException(String.format("Post pelo identificador %s não encontrado para ser atualizado.", dto.getPostId())
 		));
 		
-		this.userService.findById(UUID.fromString(dto.getUserId())).orElseThrow(() -> 
+		UserDTO userDTO = this.userService.findById(UUID.fromString(dto.getUserId())).orElseThrow(() -> 
 			new ResourceNotFoundException(String.format("Usuário pelo identificador %s não encontrado.", dto.getUserId())
-		));		
+		));	
 		
+		if(!UserUtils.getUsernameLogado().equals(userDTO.getUsername()))
+			throw new BusinessException("Não é possível atualizar comentário do post utilizando outro usuário.");
+	
 		Comment entity = this.mapper.commentRequestDTOToComment(dto);
 		entity.setPost(postMapper.postDTOToPost(postDTO));
 		entity.setDateUpdate(LocalDateTime.now());
@@ -92,8 +106,11 @@ public class CommentService {
 	
 	
 	public void deleteById(@NotNull UUID id) {
-		this.repository.findById(id).orElseThrow(
+		Comment comment = this.repository.findById(id).orElseThrow(
 				() -> new ResourceNotFoundException(String.format("Comment pelo id %s não existe.", id)));
+		
+		if(!UserUtils.getUsernameLogado().equals(comment.getUser().getUsername()) && !comment.getUser().getRole().equals(Role.ADMIN))
+			throw new BusinessException("Não é possível excluir comentário do post utilizando outro usuário, a menos que seja o administrador.");
 		
 		this.repository.deleteById(id);
 		log.info("Comment pelo id {} foi excluído com sucesso.", id);
